Implement verification flow with hardcoded role id
This is a big change that completes the verification flow and refactors
a whole bunch of stuff. This commit:

- Refactors both the server and the bot to use the new db and config
modules
- Moves the session logic to start on a bot command, instead of the bot
making a request to the server, since otherwise anyone would be able to
forge a session and spam emails
- Implements the maintenance loop to remove expired users and add the
role to verified users
- Cleans up some code and templates

There is still some work remaining around styling the pages and
implementing the remaining TODOs. Most notably, it still uses a
hardcoded role id and doesn't send out real emails yet.